api/serializers.py
import imp
from rest_framework import serializers
from django.contrib.auth import authenticate

from api.models import Coin, Product

class LoginSerializer(serializers.Serializer):
    username = serializers.CharField()
    password = serializers.CharField()

    def validate(self, data):
        user = authenticate(**data)
        if user:
            return user
        return serializers.ValidationError('incorrect credentials')


# product serializer
class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'


# coins serializer
class CoinSerializer(serializers.ModelSerializer):
    class Meta:
        model = Coin
        fields = '__all__'

